pzsai_backend/requirements.txt
pillow
ijson
cffi
peewee
psycopg2-binary
flask
